requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
pydantic[email]==2.5.0
python-multipart==0.0.6
streamlit==1.28.1
requests==2.31.0
pandas==2.1.3